refactor(GAME-038): extract DOM panel renderers to render/panels.ts
