+++ thoughts/shared/tickets/GAME-038-extract-panel-renderers.md
@@ -2,7 +2,7 @@
 id: GAME-038
 title: Extract DOM panel renderers out of mapRenderer.ts
 area: game, code-quality
-status: open
+status: resolved
 created: 2026-04-29
 ---
